+++ etmain/src/game/g_main.c
@@ -488,7 +488,7 @@ This must be the very first function compiled into the .q3vm file
 #if __GNUC__ >= 4
 #pragma GCC visibility push(default)
 #endif
-int vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6)
+intptr_t vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6)
 {
 #if __GNUC__ >= 4
 #pragma GCC visibility pop
