+++ src/qcommon/vm_local.h
@@ -26,7 +26,7 @@ If you have questions concerning this license or the applicable additional terms
 ===========================================================================
 */
 
-#include "../game/q_shared.h"
+#include "../shared/q_shared.h"
 #include "qcommon.h"
 
 typedef enum
@@ -137,7 +137,7 @@ struct vm_s
 	// DO NOT MOVE OR CHANGE THESE WITHOUT CHANGING THE VM_OFFSET_* DEFINES
 	// USED BY THE ASM CODE
 	int             programStack;	// the vm may be recursively entered
-	int             (*systemCall) (int *parms);
+	intptr_t		(*systemCall) (intptr_t * parms);
 
 	//------------------------------------
 
@@ -148,7 +148,7 @@ struct vm_s
 
 	// for dynamic linked modules
 	void           *dllHandle;
-	int             (QDECL * entryPoint) (int callNum, ...);
+	intptr_t        (QDECL * entryPoint) (int callNum, ...);
 
 	// for interpreted modules
 	qboolean        currentlyInterpreting;
